Fix pwfile parsing of usernames containing `:`
I have a number of devices that use their MAC address as the MQTT
username which cannot be changed.

By using `strrchr` to find the last `:` of a username/password line it
is now possible to parse and use usernames that contain a colon. A
base64 string will never contain a `:` so this should be safe in all
scenarios.

mosquitto_passwd man page says that the username can not contain a
colon so I've also removed the comment as this is no longer the case.

If compiling `WITH_TLS=no` the existing restriction of not being able to
use a `:` in the username or password remains.

Signed-off-by: Allan Randall <me@allanrandall.com>